FileChannelRepository 작성
